requirements.txt
autopep8==1.5.4
click==7.1.2
dataclasses==0.6
Flask==1.1.2
Flask-Cors==3.0.9
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.4
Pillow==8.0.1
pycodestyle==2.6.0
six==1.15.0
toml==0.10.2
torch==1.7.0
torchvision==0.8.1
typing-extensions==3.7.4.3
Werkzeug==1.0.1
